Finished adding John Ball's animations
The DefaultCharacter's skeleton is too different from the Fuse models' skeletons, so we need to add animations that are specifically adapted to the Fuse models. The other option was to use a "Humaniod" avatar animation type, which would work for both the DefaultCharacter and the Fuse models, however, the animations become rather stiff and unnatural-looking when using the Humanoid animation type.